+++ src/publisher/react-integration/interaction-api.ts
@@ -1,4 +1,3 @@
-import { getFiberFlags } from "./utils/getFiberFlags";
 import type { CoreApi } from "./core";
 import {
   Fiber,
@@ -7,12 +6,7 @@ import {
   SerializedElement,
 } from "../types";
 
-const MOUNTING = 1;
-const MOUNTED = 2;
-const UNMOUNTED = 3;
-
 export function createReactInteractionApi({
-  ReactTypeOfSideEffect,
   ReactTypeOfWork,
   getFiberIdThrows,
   getFiberById,
@@ -22,13 +16,12 @@ export function createReactInteractionApi({
   shouldFilterFiber,
   findFiberByHostInstance,
 }: CoreApi): ReactInterationApi {
-  const { Incomplete, NoFlags, Placement } = ReactTypeOfSideEffect;
   const { HostRoot, HostComponent, HostText, SuspenseComponent } =
     ReactTypeOfWork;
 
   function findAllCurrentHostFibers(id: number) {
     const fibers: Fiber[] = [];
-    const fiber = findCurrentFiberUsingSlowPathById(id);
+    const fiber = getFiberById(id);
 
     if (!fiber) {
       return fibers;
@@ -58,9 +51,9 @@ export function createReactInteractionApi({
     }
   }
 
-  function findNativeNodesForFiberID(id: number) {
+  function findNativeNodesForFiberId(id: number) {
     try {
-      let fiber = findCurrentFiberUsingSlowPathById(id);
+      let fiber = getFiberById(id);
 
       if (fiber === null) {
         return null;
@@ -86,7 +79,7 @@ export function createReactInteractionApi({
     }
   }
 
-  function getDisplayNameForFiberID(id: number) {
+  function getDisplayNameForFiberId(id: number) {
     const fiber = getFiberById(id);
     return fiber !== null ? getDisplayNameForFiber(fiber) : null;
   }
@@ -110,244 +103,6 @@ export function createReactInteractionApi({
     return fiber && getFiberIdThrows(fiber);
   }
 
-  // This function is copied from React and should be kept in sync:
-  // https://github.com/facebook/react/blob/main/packages/react-reconciler/src/ReactFiberTreeReflection.js
-  function isFiberMountedImpl(fiber: Fiber) {
-    let node = fiber;
-    let prevNode = null;
-
-    if (!fiber.alternate) {
-      // If there is no alternate, this might be a new tree that isn't inserted
-      // yet. If it is, then it will have a pending insertion effect on it.
-      if ((getFiberFlags(node) & Placement) !== NoFlags) {
-        return MOUNTING;
-      }
-
-      // This indicates an error during render.
-      if ((getFiberFlags(node) & Incomplete) !== NoFlags) {
-        return UNMOUNTED;
-      }
-
-      while (node.return) {
-        prevNode = node;
-        node = node.return;
-
-        if ((getFiberFlags(node) & Placement) !== NoFlags) {
-          return MOUNTING;
-        }
-
-        // This indicates an error during render.
-        if ((getFiberFlags(node) & Incomplete) !== NoFlags) {
-          return UNMOUNTED;
-        }
-
-        // If this node is inside of a timed out suspense subtree, we should also ignore errors/warnings.
-        const isTimedOutSuspense =
-          node.tag === SuspenseComponent && node.memoizedState !== null;
-        if (isTimedOutSuspense) {
-          // Note that this does not include errors/warnings in the Fallback tree though!
-          const primaryChildFragment = node.child;
-          const fallbackChildFragment = primaryChildFragment
-            ? primaryChildFragment.sibling
-            : null;
-
-          if (prevNode !== fallbackChildFragment) {
-            return UNMOUNTED;
-          }
-        }
-      }
-    } else {
-      while (node.return) {
-        node = node.return;
-      }
-    }
-
-    if (node.tag === HostRoot) {
-      // TODO: Check if this was a nested HostRoot when used with
-      // renderContainerIntoSubtree.
-      return MOUNTED;
-    }
-
-    // If we didn't hit the root, that means that we're in an disconnected tree
-    // that has been unmounted.
-    return UNMOUNTED;
-  }
-
-  // This function is copied from React and should be kept in sync:
-  // https://github.com/facebook/react/blob/main/packages/react-reconciler/src/ReactFiberTreeReflection.js
-  // It would be nice if we updated React to inject this function directly (vs just indirectly via findDOMNode).
-  // BEGIN copied code
-  function findCurrentFiberUsingSlowPathById(id: number) {
-    const fiber = getFiberById(id);
-
-    if (fiber === null) {
-      console.warn(`Could not find Fiber with id "${id}"`);
-      return null;
-    }
-
-    const { alternate } = fiber;
-    if (!alternate) {
-      // If there is no alternate, then we only need to check if it is mounted.
-      const state = isFiberMountedImpl(fiber);
-
-      if (state === UNMOUNTED) {
-        throw Error("Unable to find node on an unmounted component.");
-      }
-
-      if (state === MOUNTING) {
-        return null;
-      }
-
-      return fiber;
-    }
-
-    // If we have two possible branches, we'll walk backwards up to the root
-    // to see what path the root points to. On the way we may hit one of the
-    // special cases and we'll deal with them.
-    let a = fiber;
-    let b = alternate;
-    while (true) {
-      const parentA = a.return;
-      if (parentA === null) {
-        // We're at the root.
-        break;
-      }
-
-      const parentB = parentA.alternate;
-      if (parentB === null) {
-        // There is no alternate. This is an unusual case. Currently, it only
-        // happens when a Suspense component is hidden. An extra fragment fiber
-        // is inserted in between the Suspense fiber and its children. Skip
-        // over this extra fragment fiber and proceed to the next parent.
-        const nextParent = parentA.return;
-        if (nextParent !== null) {
-          a = b = nextParent;
-          continue;
-        }
-        // If there's no parent, we're at the root.
-        break;
-      }
-
-      // If both copies of the parent fiber point to the same child, we can
-      // assume that the child is current. This happens when we bailout on low
-      // priority: the bailed out fiber's child reuses the current child.
-      if (parentA.child === parentB.child) {
-        let child = parentA.child;
-        while (child) {
-          if (child === a) {
-            // We've determined that A is the current branch.
-            if (isFiberMountedImpl(parentA) !== MOUNTED) {
-              throw Error("Unable to find node on an unmounted component.");
-            }
-
-            return fiber;
-          }
-
-          if (child === b) {
-            // We've determined that B is the current branch.
-            if (isFiberMountedImpl(parentA) !== MOUNTED) {
-              throw Error("Unable to find node on an unmounted component.");
-            }
-
-            return alternate;
-          }
-
-          child = child.sibling;
-        }
-
-        // We should never have an alternate for any mounting node. So the only
-        // way this could possibly happen is if this was unmounted, if at all.
-        throw Error("Unable to find node on an unmounted component.");
-      }
-
-      if (a.return !== b.return) {
-        // The return pointer of A and the return pointer of B point to different
-        // fibers. We assume that return pointers never criss-cross, so A must
-        // belong to the child set of A.return, and B must belong to the child
-        // set of B.return.
-        a = parentA;
-        b = parentB;
-      } else {
-        // The return pointers point to the same fiber. We'll have to use the
-        // default, slow path: scan the child sets of each parent alternate to see
-        // which child belongs to which set.
-        //
-        // Search parent A's child set
-        let didFindChild = false;
-        let child = parentA.child;
-
-        while (child) {
-          if (child === a) {
-            didFindChild = true;
-            a = parentA;
-            b = parentB;
-            break;
-          }
-
-          if (child === b) {
-            didFindChild = true;
-            b = parentA;
-            a = parentB;
-            break;
-          }
-
-          child = child.sibling;
-        }
-
-        if (!didFindChild) {
-          // Search parent B's child set
-          child = parentB.child;
-
-          while (child) {
-            if (child === a) {
-              didFindChild = true;
-              a = parentB;
-              b = parentA;
-              break;
-            }
-
-            if (child === b) {
-              didFindChild = true;
-              b = parentB;
-              a = parentA;
-              break;
-            }
-
-            child = child.sibling;
-          }
-
-          if (!didFindChild) {
-            throw Error(
-              "Child was not found in either parent set. This indicates a bug " +
-                "in React related to the return pointer. Please file an issue."
-            );
-          }
-        }
-      }
-
-      if (a.alternate !== b) {
-        throw Error(
-          "Return fibers should always be each others' alternates. " +
-            "This error is likely caused by a bug in React. Please file an issue."
-        );
-      }
-    }
-    // If the root is not a host container, we're in a disconnected tree. I.e.
-    // unmounted.
-    if (a.tag !== HostRoot) {
-      throw Error("Unable to find node on an unmounted component.");
-    }
-
-    if (a.stateNode.current === a) {
-      // We've determined that A is the current branch.
-      return fiber;
-    }
-
-    // Otherwise B has to be current branch.
-    return alternate;
-  }
-  // END copied code
-
   function fiberToSerializedElement(fiber: Fiber): SerializedElement {
     return {
       displayName: getDisplayNameForFiber(fiber) || "Anonymous",
@@ -358,7 +113,7 @@ export function createReactInteractionApi({
   }
 
   function getOwnersList(id: number) {
-    const fiber = findCurrentFiberUsingSlowPathById(id);
+    const fiber = getFiberById(id);
 
     if (fiber === null) {
       return null;
@@ -429,9 +184,9 @@ export function createReactInteractionApi({
   }
 
   return {
-    findNativeNodesForFiberID,
-    getDisplayNameForFiberID,
-    getFiberIDForNative,
+    findNativeNodesForFiberId,
+    getDisplayNameForFiberId,
+    getFiberIdForNative: getFiberIDForNative,
     getOwnersList,
     getPathForElement,
   };