+++ src/publisher/types.ts
@@ -4,6 +4,7 @@ import {
   TransferCallTrace,
   TransferFiberChanges,
 } from "common-types";
+import { HighlightState } from "rempl";
 import {
   ExposedLeaksStateSubscription,
   TrackingObjectWeakRef,
@@ -462,12 +463,12 @@ export type ReactDevtoolsHookHandlers = {
   handleCommitFiberUnmount: (fiber: Fiber) => void;
 };
 export type ReactInterationApi = {
-  findNativeNodesForFiberID: (id: number) => any[] | null;
-  getFiberIDForNative: (
+  findNativeNodesForFiberId: (id: number) => any[] | null;
+  getFiberIdForNative: (
     hostInstance: NativeType,
     findNearestUnfilteredAncestor?: boolean
   ) => number | null;
-  getDisplayNameForFiberID: (id: number) => string | null;
+  getDisplayNameForFiberId: (id: number) => string | null;
   getOwnersList: (id: number) => Array<SerializedElement> | null;
   getPathForElement: (id: number) => Array<PathFrame> | null;
 };
@@ -499,10 +500,20 @@ export type ReactDispatcherTrapApi = {
   flushDispatchCalls: (root: FiberRoot) => FiberDispatchCall[];
 };
 
+export type HighlightApi = {
+  subscribe: (fn: (state: HighlightState) => void) => void;
+  startHighlight: (fiberId: number) => void;
+  stopHighlight: () => void;
+  startInspect: () => void;
+  stopInspect: () => void;
+};
+
 export type RemoteCommandsApi = Omit<
   MemoryLeakDetectionApi,
   "getLeakedObjectsProbe"
->;
+> & {
+  highlightApi: HighlightApi;
+};
 
 export type ReactIntegrationApi = ReactDevtoolsHookHandlers &
   ReactInterationApi &