+++ src/ui/utils/selection.tsx
@@ -185,3 +185,4 @@ export const useSelectedId = () => {
 
   return { selectedId: state, select };
 };
+